Move Binary to Encoder
